test: some of the test in signup and login are updated
