+++ frontend/tests/fixtures/constant.ts
@@ -1,0 +1,1 @@
+export const version = 'v1.0.0';
